feat[OK-238] Add board name in response
