+++ server/src/shared/interfaces/response-objects/CreatePermissionsStr.ts
@@ -1,0 +1,5 @@
+export interface CreatePermissionStrResponse {
+  permissionStr: string;
+  ttlInMs: number;
+  permission: string;
+}
