+++ server/src/shared/interfaces/response-objects/GrantPermission.ts
@@ -1,0 +1,5 @@
+export interface GrantPermissionResponse {
+  boardId: string;
+  name: string;
+  permission: string;
+}
